refactor(structure): relocated file structure
