+++ AI/ocr/.env
@@ -1,0 +1,11 @@
+# YOLO model
+YOLO_MODEL_PATH=ocr/best.pt
+
+# OCR Config
+OCR_CONFIG_NAME=vgg_transformer
+OCR_DEVICE=cuda
+CONFIDENCE_THRESHOLD=0.25
+
+# Server Config
+HOST=0.0.0.0
+PORT=8000
