--- AI/ocr/README.md
@@ -1,1 +1,0 @@
-uvicorn app.main:app --reload
