+++ AI/ocr/config.py
