+++ AI/ocr/main_ocr.py
@@ -1,10 +1,10 @@
 from fastapi import FastAPI, UploadFile, HTTPException
 from PIL import Image
-from app.models import OCRResult
-from app.utils.image_utils import crop_image, preprocess_image
-from app.utils.ocr_utils import clean_text
-from app.utils.yolo_utils import load_yolo_model, get_label_map
-from app.config import load_ocr_model
+from ocr.models import OCRResult
+from ocr.utils.image_utils import crop_image, preprocess_image
+from ocr.utils.ocr_utils import clean_text
+from ocr.utils.yolo_utils import load_yolo_model, get_label_map
+from ocr.config import load_ocr_model
 import numpy as np
 import json
