--- AI/ocr/requirements.txt
@@ -1,8 +1,0 @@
-fastapi
-uvicorn
-ultralytics
-opencv-python
-Pillow
-matplotlib
-vietocr
-numpy
