+++ AI/ocr/utils/image_utils.py
