+++ AI/ocr/utils/ocr_utils.py
